chore: move translator functions into separate modules
